cmd/snap-update-ns, osutil: support AllowSocketFile in snap-update-ns entries
For the cups interface, it is desirable to allow using the mount backend to
mount a socket file from one mount namespace to another, however emitting such
rules from the mount backend into the fstab that snap-update-ns does not permit
socket files, as it checks if the file is regular.

Allow mount entries to specify that the file being mounted can be a socket
file under the specific circumstances where the mount backend specifies an
option for the mount entry as such.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>