interfaces/cups: add cups-socket attr, use to specify mount rules in backend
Make the cups interface slot have a mandatory attribute, cups-socket, which is
used to specify the location of the cups socket that the slotting snap is
going to share with client snaps.

This socket file then is mounted into client snap's mount namespaces via the
mount backend at /run/cups/cups.sock, such that client applications wishing to
print don't need to adjust the location of where they expect the cups socket to
be at all, but still will end up sending print requests to the cups snap, which
will mediate requests to print based on whether or not the client snap has a
connected cups plug or not.

The primary advantage of this is that it means we don't need to make the cups
interface implictly slotted by the system snap, and it can always be provided
by the cups snap, and we can then make any snap with a cups interface plug
auto-connect to the slot from the cups snap unambiguously, providing all snap
clients the ability to print without any extra connections necessary,
regardless of their distro or what the client snap is.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>