store/store_download.go: use system snap provided xdelta3 priority + fallback
Prefer using xdelta3 from the system snap if it is available, and if it is not
available or it fails, fallback to using the host version to apply a delta.

This has the effect that on older systems like Centos 7 and Amazon Linux 2,
where for example xdelta3 from the distro does not support LZMA, when a delta
generated using LZMA is encountered, we will first try the system snap provided
xdelta3 which should succeed. But on the off chance that it doesn't succeed,
but xdelta3 still exists in the system snap, we will retry the execution using
an xdelta3 from the system path. This would help with situations like where the
delta generated by the store moves further forward, using a new compression
algorithm that is not supported in xenial for example and someone is using an
older core snap but a newer distro that supports it.

This commit is missing tests in order to get feedback on the implementation.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>